requirements.txt
nonebot[scheduler]~=1.6.0
base58==2.0.0
pycryptodome~=3.9.7